Oprava vymazavania a pridavania formularu

+++ backend/app.js
@@ -433,9 +433,18 @@ app.post('/forms/create', (req,res)=>{
     type = req.body.type;
     reason = req.body.reason;
     details = req.body.details;
-    if(reason == undefined || dog_id == undefined || type==undefined){
+    if(details == undefined || dog_id == undefined || type==undefined){
         res.status(404).json({"message": "Not all attributes provided"});
         return;
+    }
+
+    if (reason == undefined){
+        if (type == 1) {
+            res.status(404).json({"message": "Not all attributes provided"});
+            return;
+        }
+        if (type == 2)
+            reason = null;
     }
 
     if (shelter == true) {
@@ -448,12 +457,12 @@ app.post('/forms/create', (req,res)=>{
         return;
     }
 
-    db.one("INSERT INTO forms(form_type,reason,details,dog_id,user_id,created_at,finished) VALUES ($1, $2, $3,$4, CURRENT_DATE, false) RETURNING ID", [type,reason,details,dog_id,userID])
+    db.one("INSERT INTO forms(form_type,reason,details,dog_id,user_id,created_at,finished) VALUES ($1, $2, $3, $4, $5, CURRENT_DATE, false) RETURNING ID", [type,reason,details,dog_id,userID])
     .then((data)=>{
         //ak je to vencenie treba este sparovat termin s formularom
         if(type==2){
             term_id = req.body.term_id;
-            if(term_id == undefined){
+            if(term_id == undefined || term_id == null){
                 db.any("DELETE FROM forms WHERE id=$1",[data.id])
                 .then((data)=>{
                     res.status(400).json({"message": "Not all attributes provided"});
@@ -462,15 +471,16 @@ app.post('/forms/create', (req,res)=>{
                     res.status(404).json({"message": "No data to be deleted and not all attributes provided"});
                 }) 
             }
-            db.one("UPDATE terms SET form_id=$1 WHERE terms.id=$2 RETURNING id",[data.id,term_id])
+            db.one("UPDATE terms SET form_id=$1, user_id=$2, free=false WHERE terms.id=$3 RETURNING id",[data.id,userID,term_id])
             .then((data)=>{
-                res.status(200).json({"message":"OK"})
+                res.status(200).json({"message":"OK"});
             })
             .catch((error)=>{
                 res.status(404).json({"message": "No data to be updated"});
             })
         }
-        res.status(200).json({"message": "OK"});
+        else
+            res.status(200).json({"message": "OK"});
     })
     .catch((error)=>{
         res.status(400).json({"message": "No data found"});
@@ -590,11 +600,18 @@ app.delete('/forms/delete',(req,res)=>{
     }
 
     userID = tokens[token]["id"];
-    db.one("DELETE from forms WHERE id=$2 AND (forms.user_id=$1 OR forms.dog_id IN (SELECT dogs.id FROM dogs WHERE shelter_id = $1)) RETURNING dog_id",[userID,form_id])
+    db.one("DELETE from forms WHERE id=$2 AND (forms.user_id=$1 OR forms.dog_id IN (SELECT dogs.id FROM dogs WHERE shelter_id = $1)) RETURNING dog_id, finished",[userID,form_id])
     .then((data)=>{
-        db.any("DELETE FROM terms WHERE dog_id = $1 and form_id = $2", [data.dog_id, form_id])
-        .then((data) => res.status(200).json({"message":"OK"}))
-        .catch((error)=> res.status(404).json({"message":"No data to be deleted"}))
+        if (data.finished == true) {
+            db.any("DELETE FROM terms WHERE dog_id = $1 and form_id = $2", [data.dog_id, form_id])
+            .then((data) => res.status(200).json({"message":"OK"}))
+            .catch((error)=> res.status(404).json({"message":"No data to be deleted"}))
+        }
+        else {
+            db.any("UPDATE terms SET form_id = null, user_id = null, free = true WHERE dog_id = $1 and form_id = $2", [data.dog_id, form_id])
+            .then((data) => res.status(200).json({"message":"OK"}))
+            .catch((error)=> res.status(404).json({"message":"No data to be updated"}))
+        }       
     })
     .catch((error)=>{
         res.status(404).json({"message": "No data to be deleted"});
